Pixed Unit Satue and add more skills freure
Pixed Unit Satue and add more skills freure

+++ Assets/Scripts/Skills/Core/BuffData.cs
@@ -8,8 +8,7 @@ public class BuffData : ScriptableObject
     public int duration;
 
     public int attackBonus;
-    public int defenseBonus;
-
+    public float defenseBonus;
     public int speedBonus;
 
     public int hitBonus;
